Add note preview and autoplay options to instrument editor
