+++ tools/tracker/src/services/PatternService.js
@@ -1,6 +1,16 @@
 import { patterns, patternsMeta, patternMuteStates, sequence } from '../stores.js';
 import { clamp, round } from '../lib/utils.js';
 
+const NOTE_NAMES = [
+  'REL', '---',
+  'C-1', 'C#1', 'D-1', 'D#1', 'E-1', 'F-1',
+  'F#1', 'G-1', 'G#1', 'A-1', 'A#1', 'B-1',
+  'C-2', 'C#2', 'D-2', 'D#2', 'E-2', 'F-2',
+  'F#2', 'G-2', 'G#2', 'A-2', 'A#2', 'B-2',
+  'C-3', 'C#3', 'D-3', 'D#3', 'E-3', 'F-3',
+  'F#3', 'G-3', 'G#3', 'A-3', 'A#3', 'B-3'
+];
+
 // This is used to auto-name patterns
 let patternNo = 0;
 
@@ -237,3 +247,12 @@ export const setNote = (pattern, channel, row, note) => {
     return patterns;
   });
 };
+
+
+/**
+ * Returns a human-readable string for a note ID.
+ *
+ * @param {number} noteId - The note
+ * @returns {string} The name of the note (e.g. `C#1`)
+ */
+export const getNoteName = noteId => NOTE_NAMES[noteId];
